+++ DumbRequestManager/Classes/Map.cs
@@ -1,0 +1,6 @@
+﻿namespace DumbRequestManager.Classes;
+
+public class Map
+{
+    
+}