+++ DumbRequestManager/Classes/QueuedMap.cs
@@ -161,7 +161,11 @@ public class NoncontextualizedDifficulty
     };
     public NoncontextualizedDifficulty(BeatmapLevel level, BeatmapKey key, BeatmapBasicData mapData, CachedMap? cachedDetails)
     {
+#if V1_44_0
         SongData.DifficultyData? diffData = SongCore.Collections.GetCustomLevelSongDifficultyData(key);
+#else
+        ExtraSongData.DifficultyData? diffData = SongCore.Collections.RetrieveDifficultyData(level, key);
+#endif
         
         int mods = 0;
         if (diffData != null)
@@ -331,6 +335,8 @@ public class NoncontextualizedSong
     [JsonProperty] public bool Blacklisted => BlacklistManager.ContainsKey(BsrKey);
 
     [JsonProperty] public NoncontextualizedDifficulty[] Diffs { get; set; } = [];
+    
+    [JsonProperty] public string[]? Tags { get; set; } = [];
 
     private CoverImageContainer _coverImageContainer = null!;
     public byte[]? CoverImage => _coverImageContainer.CoverImage ?? null;
@@ -382,7 +388,7 @@ public class NoncontextualizedSong
         MetadataHasSplicedCensor = Censor.Check([Title, SubTitle, Artist, Mapper]);
         Duration = song.Duration;
         Votes = [song.Votes.Up, song.Votes.Down];
-        Rating = CalculateRating(song.Votes.Up, song.Votes.Down);
+        Rating = (float)song.Votes.Score;
         UploadTime = song.UploadTimestamp;
         LastUpdated = song.LastUpdateTimestamp;
         Cover = $"https://cdn.beatsaver.com/{song.Hash}.jpg";
@@ -395,6 +401,8 @@ public class NoncontextualizedSong
         UsesMappingExtensions = ((CachedMapMods)song.Mods).HasFlag(CachedMapMods.MappingExtensions);
         UsesNoodleExtensions = ((CachedMapMods)song.Mods).HasFlag(CachedMapMods.NoodleExtensions);
         UsesVivify = ((CachedMapMods)song.Mods).HasFlag(CachedMapMods.Vivify);
+        Tags = song.Tags ?? [];
+        
         // (Local cache doesn't cache automapped maps, always false)
         Diffs = song.Difficulties.Select(x => new NoncontextualizedDifficulty(x, song)).ToArray();
         if (!skipCoverImage)
@@ -434,7 +442,10 @@ public class NoncontextualizedSong
         UsesCinema = song.LatestVersion.Difficulties.Any(x => x.Cinema);
         UsesMappingExtensions = song.LatestVersion.Difficulties.Any(x => x.MappingExtensions);
         UsesNoodleExtensions = song.LatestVersion.Difficulties.Any(x => x.NoodleExtensions);
+        UsesVivify = song.LatestVersion.Difficulties.Any(x => x.Vivify);
         Diffs = song.LatestVersion.Difficulties.Select(x => new NoncontextualizedDifficulty(x)).ToArray();
+        Tags = song.Tags.ToArray() ?? [];
+        
         if (!skipCoverImage)
         {
             _coverImageContainer = new CoverImageContainer(song);
@@ -465,8 +476,12 @@ public class NoncontextualizedSong
         UsesVivify = false;
         foreach (BeatmapKey key in level.GetBeatmapKeys())
         {
+#if V1_44_0
             SongData.DifficultyData? diffData = SongCore.Collections.GetCustomLevelSongDifficultyData(key);
-
+#else
+            ExtraSongData.DifficultyData? diffData = SongCore.Collections.RetrieveDifficultyData(level, key);
+#endif
+            
             if (diffData == null)
             {
                 continue;
@@ -488,7 +503,7 @@ public class NoncontextualizedSong
 
             BsrKey = cachedDetails.Key.ToString("x") ?? string.Empty;
             Votes = [cachedDetails.Votes.Up, cachedDetails.Votes.Down];
-            Rating = CalculateRating(cachedDetails.Votes.Up, cachedDetails.Votes.Down);
+            Rating = (float)cachedDetails.Votes.Score;
             LastUpdated = cachedDetails.LastUpdateTimestamp;
             UploadTime = cachedDetails.UploadTimestamp;
             Cover = $"https://cdn.beatsaver.com/{cachedDetails.Hash}.jpg";
@@ -496,6 +511,7 @@ public class NoncontextualizedSong
             ScoreSaberRanked = cachedDetails.Difficulties.Any(x => x.RankedStatus.ScoreSaber.IsRanked);
             Curated = cachedDetails.Curator != null;
             CuratorName = cachedDetails.Curator ?? string.Empty;
+            Tags = cachedDetails.Tags;
         }
         else
         {
@@ -514,6 +530,7 @@ public class NoncontextualizedSong
                 ScoreSaberRanked = beatsaverDetails.Ranked;
                 Curated = beatsaverDetails.BeatmapCurator != null;
                 CuratorName = beatsaverDetails.BeatmapCurator?.Name ?? string.Empty;
+                Tags = beatsaverDetails.Tags.ToArray();
             }
         }